We do not need to set key ranges in tipb.SelectRequest.
We only need kv.KeyRange and kvproto/coprocessor.KeyRange.
This run faster in some of our cases. I will make code more cleaner latter in this PR.

@coocood @BusyJay @tiancaiamao PTAL