@shenli 2016-09-04
Reduce make slice when running EncodeRowKeyWithHandle.

Before:
BenchmarkEncodeRowKeyWithHandle-4	10000000	       118 ns/op
ok  	github.com/pingcap/tidb/tablecodec	1.324s

After:
BenchmarkEncodeRowKeyWithHandle-4	30000000	        47.7 ns/op
ok  	github.com/pingcap/tidb/tablecodec	1.498s