@shenli 2016-09-04
BenchmarkEncodeEndKey-4             	20000000	        97.2 ns/op
BenchmarkEncodeRowKeyWithPrefixNex-4 	10000000	       121 ns/op